homework-3: changed (int radius) to (double radius)

+++ src/hw3/Circle.java
@@ -1,14 +1,14 @@
 package hw3;
 
 public class Circle {
-    private int radius;
+    private double radius;
 
     public void calculationCircleArea() {
-        int s = (int) (Math.PI * Math.pow(radius, 2));
+        double s = (Math.PI * Math.pow(radius, 2));
         System.out.println("Circle area is: " + s + " cm2");
     }
 
-    public Circle(int radius) {
+    public Circle(double radius) {
         this.radius = radius;
     }
 }
